accounts: refactor API for generalized USB wallets
